make animated slider code into custom widgets for improved modularity and readability.
